move 'innerJoin' tests to separate file
