+++ types/ramda/tsconfig.json
@@ -99,6 +99,7 @@
         "test/includes-tests.ts",
         "test/indexBy-tests.ts",
         "test/indexOf-tests.ts",
-        "test/init-tests.ts"
+        "test/init-tests.ts",
+        "test/innerJoin-tests.ts"
     ]
 }
